Move horner functions into its own file
