+++ src/aberth.rs
@@ -1,84 +1,11 @@
 #![allow(non_snake_case)]
 
+use super::horner::{horner_eval_c, horner_eval_f};
 use super::Options;
 use num::Complex;
 // use lds_rs::lds::Circle;
 
 const TWO_PI: f64 = std::f64::consts::TAU;
-
-/// Horner evalution (float)
-/// 
-/// The `horner_eval_f` function in Rust implements the Horner's method for evaluating a polynomial with
-/// given coefficients at a specific value.
-/// 
-/// Arguments:
-/// 
-/// * `coeffs`: A vector of floating-point coefficients representing a polynomial. The coefficients are
-/// ordered from highest degree to lowest degree. For example, the polynomial 10x^8 + 34x^7 + 75x^6 +
-/// 94x^5 + 150x^4 + 94x^
-/// * `zval`: The `zval` parameter in the `horner_eval_f` function represents the value at which the
-/// polynomial is evaluated. It is of type `f64`, which means it is a floating-point number.
-/// 
-/// Returns:
-/// 
-/// The function `horner_eval_f` returns a `f64` value, which is the result of evaluating the polynomial
-/// with the given coefficients at the specified value `zval`.
-///
-/// # Examples:
-///
-/// ```
-/// use bairstow::aberth::horner_eval_f;
-/// use approx_eq::assert_approx_eq;
-///
-/// let coeffs = vec![10.0, 34.0, 75.0, 94.0, 150.0, 94.0, 75.0, 34.0, 10.0];
-/// let px = horner_eval_f(&coeffs, 2.0);
-///
-/// assert_approx_eq!(px, 18250.0);
-/// ```
-pub fn horner_eval_f(coeffs: &[f64], zval: f64) -> f64 {
-    coeffs
-        .iter()
-        .copied()
-        .reduce(|res, coeff| res * zval + coeff)
-        .unwrap()
-}
-
-/// Horner evalution (complex)
-/// 
-/// The `horner_eval_c` function in Rust implements the Horner evaluation method for complex
-/// polynomials.
-/// 
-/// Arguments:
-/// 
-/// * `coeffs`: A vector of coefficients representing a polynomial. The coefficients are in descending
-/// order of degree. For example, the polynomial 10x^8 + 34x^7 + 75x^6 + 94x^5 + 150x^4 + 94x^3 + 75
-/// * `zval`: The `zval` parameter is a complex number that represents the value at which the polynomial
-/// is evaluated.
-/// 
-/// Returns:
-/// 
-/// The function `horner_eval_c` returns a complex number of type `Complex<f64>`.
-///
-/// # Examples:
-///
-/// ```
-/// use bairstow::aberth::horner_eval_c;
-/// use approx_eq::assert_approx_eq;
-/// use num::Complex;
-///
-/// let coeffs = vec![10.0, 34.0, 75.0, 94.0, 150.0, 94.0, 75.0, 34.0, 10.0];
-/// let px = horner_eval_c(&coeffs, &Complex::new(1.0, 2.0));
-///
-/// assert_approx_eq!(px.re, 6080.0);
-/// assert_approx_eq!(px.im, 9120.0);
-/// ```
-pub fn horner_eval_c(coeffs: &[f64], zval: &Complex<f64>) -> Complex<f64> {
-    coeffs
-        .iter()
-        .map(|coeff| Complex::<f64>::new(*coeff, 0.0))
-        .reduce(|res, coeff| res * zval + coeff)
-        .unwrap()
-}
 
 /// Initial guess for Aberth's method
 ///
